+++ .buildkite/gen-pipeline.sh
@@ -99,32 +99,18 @@ run_mpi_pytest() {
   local oneccl_env=${3:-}
   oneccl_env=$(echo ${oneccl_env//:/ })
 
-  local exclude_keras=""
-  if [[ ${test} == *"tf2_"* ]] || [[ ${test} == *"tfhead"* ]]; then
-    # TODO: support for Keras + TF 2.0 and TF-Keras 2.0
-    exclude_keras="| sed 's/test_keras.py//g' | sed 's/test_tensorflow_keras.py//g'"
-  else
-    exclude_keras="| sed 's/[a-z_]*tensorflow2[a-z_.]*//g'"
-  fi
-
-  local excluded_tests="| sed 's/test_interactiverun.py//g' | sed 's/test_spark_keras.py//g' | sed 's/test_spark_torch.py//g'"
-
-  # Spark and Run test does not need to be executed with horovodrun, but we still run it below.
-  local exclude_standalone_test="| sed 's/test_spark.py//g' | sed 's/test_run.py//g' | sed 's/test_ray.py//g' | sed 's/test_ray_elastic.py//g'"
-  local standalone_tests="test_spark.py test_run.py"
-
   # pytests have 4x GPU use cases and require a separate queue
   run_test "${test}" "${queue}" \
-    ":pytest: Run PyTests (${test})" \
-    "bash -c \"${oneccl_env} cd /horovod/test && (ls -1 test_*.py ${exclude_keras} ${excluded_tests} ${exclude_standalone_test} | xargs -n 1 \\\$(cat /mpirun_command) /bin/bash /pytest.sh mpi)\"" \
+    ":pytest: MPI Parallel PyTests (${test})" \
+    "bash -c \"${oneccl_env} cd /horovod/test/parallel && (ls -1 test_*.py | xargs -n 1 \\\$(cat /mpirun_command) /bin/bash /pytest.sh mpi)\"" \
     5
   run_test "${test}" "${queue}" \
-    ":pytest: Run PyTests Standalone (${test})" \
-    "bash -c \"${oneccl_env} cd /horovod/test && pytest --forked -v --capture=fd --continue-on-collection-errors --junit-xml=/artifacts/junit.mpi.standalone.xml ${standalone_tests}\"" \
-    5
+    ":pytest: MPI Single PyTests (${test})" \
+    "bash -c \"${oneccl_env} cd /horovod/test/single && (ls -1 test_*.py | xargs -n 1 /bin/bash /pytest_standalone.sh mpi)\"" \
+    10
 
   run_test "${test}" "${queue}" \
-    ":pytest: Run Cluster PyTests (${test})" \
+    ":pytest: MPI Cluster PyTests (${test})" \
     "bash -c \"${oneccl_env} /etc/init.d/ssh start && cd /horovod/test/integration && pytest --forked -v --capture=fd --continue-on-collection-errors --junit-xml=/artifacts/junit.mpi.static.xml test_static_run.py\""
 }
 
@@ -139,35 +125,35 @@ run_mpi_integration() {
     # TODO: support mpich
     run_test "${test}" "${queue}" \
       ":jupyter: Run PyTests test_interactiverun (${test})" \
-      "bash -c \"cd /horovod/test && pytest -v --capture=no --continue-on-collection-errors --junit-xml=/artifacts/junit.mpi.integration.xml test_interactiverun.py\""
+      "bash -c \"cd /horovod/test && pytest -v --capture=no --continue-on-collection-errors --junit-xml=/artifacts/junit.mpi.integration.xml integration/test_interactiverun.py\""
   fi
 
   # Legacy TensorFlow tests
   if [[ ${test} != *"tf2_"* ]] && [[ ${test} != *"tfhead"* ]]; then
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow MNIST (${test})" \
+      ":tensorflow: MPI TensorFlow MNIST (${test})" \
       "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/examples/tensorflow/tensorflow_mnist.py\""
 
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow Eager MNIST (${test})" \
+      ":tensorflow: MPI TensorFlow Eager MNIST (${test})" \
       "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/examples/tensorflow/tensorflow_mnist_eager.py\""
 
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test Keras MNIST (${test})" \
+      ":tensorflow: MPI Keras MNIST (${test})" \
       "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/examples/keras/keras_mnist_advanced.py\""
 
     run_test "${test}" "${queue}" \
-      ":fire: Test PyTorch MNIST (${test})" \
+      ":fire: MPI PyTorch MNIST (${test})" \
       "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/examples/pytorch/pytorch_mnist.py\""
   fi
 
   if [[ ${test} == *"mxnet2_"* ]] || [[ ${test} == *"mxnethead"* ]]; then
       run_test "${test}" "${queue}" \
-               ":muscle: Test MXNet2 MNIST (${test})" \
+               ":muscle: MPI MXNet2 MNIST (${test})" \
                "bash -c \"${oneccl_env} OMP_NUM_THREADS=1 \\\$(cat /mpirun_command) python /horovod/examples/mxnet/mxnet2_mnist.py\""
   else
       run_test "${test}" "${queue}" \
-               ":muscle: Test MXNet MNIST (${test})" \
+               ":muscle: MPI MXNet MNIST (${test})" \
                "bash -c \"${oneccl_env} OMP_NUM_THREADS=1 \\\$(cat /mpirun_command) python /horovod/examples/mxnet/mxnet_mnist.py\""
   fi
 
@@ -175,15 +161,15 @@ run_mpi_integration() {
   # a framework-specific functionality
   if [[ ${test} == *"tf1_15_0"* ]]; then
     run_test "${test}" "${queue}" \
-      ":muscle: Test Stall (${test})" \
-      "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/test/test_stall.py\""
+      ":muscle: MPI Stall (${test})" \
+      "bash -c \"${oneccl_env} \\\$(cat /mpirun_command) python /horovod/test/integration/test_stall.py\""
 
     if [[ ${test} == *"openmpi"* ]]; then
       run_test "${test}" "${queue}" \
-        ":terminal: Test Horovodrun (${test})" \
+        ":terminal: MPI Horovodrun (${test})" \
         "horovodrun -np 2 -H localhost:2 python /horovod/examples/tensorflow/tensorflow_mnist.py"
       run_test "${test}" "${queue}" \
-        ":terminal: Test Horovodrun (${test})" \
+        ":terminal: MPI Horovodrun (${test})" \
         "bash -c \"echo 'localhost slots=2' > hostfile && horovodrun -np 2 -hostfile hostfile python /horovod/examples/mxnet/mxnet_mnist.py\""
     fi
   fi
@@ -191,11 +177,11 @@ run_mpi_integration() {
   # TensorFlow 2.0 tests
   if [[ ${test} == *"tf2_"* ]] || [[ ${test} == *"tfhead"* ]]; then
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow 2.0 MNIST (${test})" \
+      ":tensorflow: MPI TensorFlow 2.0 MNIST (${test})" \
       "bash -c \"\\\$(cat /mpirun_command) python /horovod/examples/tensorflow2/tensorflow2_mnist.py\""
 
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow 2.0 Keras MNIST (${test})" \
+      ":tensorflow: MPI TensorFlow 2.0 Keras MNIST (${test})" \
       "bash -c \"\\\$(cat /mpirun_command) python /horovod/examples/tensorflow2/tensorflow2_keras_mnist.py\""
   fi
 }
@@ -213,33 +199,17 @@ run_gloo_pytest() {
   local test=$1
   local queue=$2
 
-  local exclude_keras=""
-  if [[ ${test} == *"tf2_"* ]] || [[ ${test} == *"tfhead"* ]]; then
-    # TODO: support for Keras + TF 2.0 and TF-Keras 2.0
-    exclude_keras="| sed 's/test_keras.py//g' | sed 's/test_tensorflow_keras.py//g'"
-  else
-    exclude_keras="| sed 's/[a-z_]*tensorflow2[a-z_.]*//g'"
-  fi
-
-  # These are tested as integration style tests.
-  local excluded_tests="| sed 's/test_interactiverun.py//g' | sed 's/test_spark_keras.py//g' | sed 's/test_spark_torch.py//g'"
-
-  # Spark and Run test does not need to be executed with horovodrun, but we still run it below.
-  local exclude_standalone_test="| sed 's/test_spark.py//g' | sed 's/test_run.py//g' | sed 's/test_ray.py//g' | sed 's/test_ray_elastic.py//g'"
-  local standalone_tests="test_spark.py test_run.py"
-  local standalone_ray_tests="test_ray.py test_ray_elastic.py"
-
   run_test "${test}" "${queue}" \
-    ":pytest: Run PyTests (${test})" \
-    "bash -c \"cd /horovod/test && (ls -1 test_*.py ${exclude_keras} ${excluded_tests} ${exclude_standalone_test} | xargs -n 1 horovodrun -np 2 -H localhost:2 --gloo /bin/bash /pytest.sh gloo)\"" \
+    ":pytest: Gloo Parallel PyTests (${test})" \
+    "bash -c \"cd /horovod/test/parallel && (ls -1 test_*.py | xargs -n 1 horovodrun -np 2 -H localhost:2 --gloo /bin/bash /pytest.sh gloo)\"" \
     5
   run_test "${test}" "${queue}" \
-    ":pytest: Run PyTests Standalone (${test})" \
-    "bash -c \"cd /horovod/test && pytest --forked -v --capture=fd --continue-on-collection-errors --junit-xml=/artifacts/junit.gloo.standalone.xml ${standalone_tests} && pytest --forked -v --capture=fd --continue-on-collection-errors --junit-xml=/artifacts/junit.gloo.standalone.xml ${standalone_ray_tests}\"" \
+    ":pytest: Gloo Single PyTests (${test})" \
+    "bash -c \"cd /horovod/test/single && (ls -1 test_*.py | xargs -n 1 /bin/bash /pytest_standalone.sh gloo)\"" \
     10
 
   run_test "${test}" "${queue}" \
-    ":pytest: Run Cluster PyTests (${test})" \
+    ":pytest: Gloo Cluster PyTests (${test})" \
     "bash -c \"/etc/init.d/ssh start && cd /horovod/test/integration && pytest --forked -v --capture=fd --continue-on-collection-errors --junit-xml=/artifacts/junit.gloo.static.xml test_static_run.py\""
 }
 
@@ -250,33 +220,33 @@ run_gloo_integration() {
   # TensorFlow 2.0 tests
   if [[ ${test} == *"tf2_"* ]] || [[ ${test} == *"tfhead"* ]]; then
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow 2.0 MNIST (${test})" \
+      ":tensorflow: Gloo TensorFlow 2.0 MNIST (${test})" \
       "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/tensorflow2/tensorflow2_mnist.py"
 
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow 2.0 Keras MNIST (${test})" \
+      ":tensorflow: Gloo TensorFlow 2.0 Keras MNIST (${test})" \
       "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/tensorflow2/tensorflow2_keras_mnist.py"
   else
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test TensorFlow MNIST (${test})" \
+      ":tensorflow: Gloo TensorFlow MNIST (${test})" \
       "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/tensorflow/tensorflow_mnist.py"
 
     run_test "${test}" "${queue}" \
-      ":tensorflow: Test Keras MNIST (${test})" \
+      ":tensorflow: Gloo Keras MNIST (${test})" \
       "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/keras/keras_mnist_advanced.py"
   fi
 
   run_test "${test}" "${queue}" \
-    ":fire: Test PyTorch MNIST (${test})" \
+    ":fire: Gloo PyTorch MNIST (${test})" \
     "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/pytorch/pytorch_mnist.py"
 
   if [[ ${test} == *"mxnet2_"* ]] || [[ ${test} == *"mxnethead"* ]]; then
       run_test "${test}" "${queue}" \
-               ":muscle: Test MXNet2 MNIST (${test})" \
+               ":muscle: Gloo MXNet2 MNIST (${test})" \
                "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/mxnet/mxnet2_mnist.py"
   else
       run_test "${test}" "${queue}" \
-               ":muscle: Test MXNet MNIST (${test})" \
+               ":muscle: Gloo MXNet MNIST (${test})" \
                "horovodrun -np 2 -H localhost:2 --gloo python /horovod/examples/mxnet/mxnet_mnist.py"
   fi
 
@@ -318,6 +288,12 @@ run_spark_integration() {
 
   # Horovod Spark Estimator tests
   if [[ ${test} != *"mpich"* && ${test} != *"oneccl"* ]]; then
+    if [[ ${queue} != *gpu* ]]; then
+      run_test "${test}" "${queue}" \
+        ":spark: Spark PyTests (${test})" \
+        "bash -c \"cd /horovod/test/integration && (ls -1 test_spark*.py | xargs -n 1 /bin/bash /pytest_standalone.sh spark)\""
+    fi
+
     if [[ ${test} != *"tf2"* && ${test} != *"tfhead"* ]]; then
       run_test "${test}" "${queue}" \
         ":spark: Spark Keras Rossmann Run (${test})" \
@@ -330,12 +306,6 @@ run_spark_integration() {
       run_test "${test}" "${queue}" \
         ":spark: Spark Keras MNIST (${test})" \
         "bash -c \"OMP_NUM_THREADS=1 python /horovod/examples/spark/keras/keras_spark_mnist.py --num-proc 2 --work-dir /work --data-dir /data --epochs 3\""
-
-      if [[ ${queue} != *gpu* ]]; then
-        run_test "${test}" "${queue}" \
-          ":spark: PyTests Spark Estimators (${test})" \
-          "bash -c \"cd /horovod/test && pytest --forked -v --capture=no --continue-on-collection-errors --junit-xml=/artifacts/junit.spark.integration.xml test_spark_keras.py test_spark_torch.py\""
-      fi
     fi
 
     run_test "${test}" "${queue}" \