+++ test/integration/test_spark_keras.py
@@ -14,9 +14,12 @@
 # ==============================================================================
 
 import collections
-from distutils.version import LooseVersion
 import logging
+import os
+import sys
 import warnings
+
+from distutils.version import LooseVersion
 
 import mock
 import numpy as np
@@ -32,6 +35,8 @@ from horovod.spark.common import constants, util
 from horovod.spark.keras import remote
 from horovod.spark.keras.estimator import EstimatorParams
 from horovod.spark.keras.util import _custom_sparse_to_dense_fn, _serialize_param_value, BareKerasUtil, TFKerasUtil
+
+sys.path.append(os.path.join(os.path.dirname(__file__), os.pardir, 'utils'))
 
 from common import temppath
 from spark_common import CallbackBackend, create_mnist_data, create_xor_data, local_store, spark_session
@@ -67,6 +72,7 @@ def get_mock_fit_fn():
     return fit
 
 
+@pytest.mark.skipif(LooseVersion(tf.__version__) >= LooseVersion('2.0.0'), reason='TensorFlow v1 tests')
 class SparkKerasTests(tf.test.TestCase):
     def __init__(self, *args, **kwargs):
         super(SparkKerasTests, self).__init__(*args, **kwargs)