+++ test/integration/test_static_run.py
@@ -31,7 +31,7 @@ from horovod.runner import _HorovodArgs
 from horovod.runner.launch import _check_all_hosts_ssh_successful, _run
 from horovod.runner.mpi_run import mpi_available, is_mpich
 
-sys.path.append(os.path.join(os.path.dirname(__file__), os.pardir))
+sys.path.append(os.path.join(os.path.dirname(__file__), os.pardir, 'utils'))
 
 from common import capture