+++ test/parallel/test_tensorflow_keras.py
@@ -13,36 +13,69 @@
 # limitations under the License.
 # ==============================================================================
 
-"""Tests for horovod.keras."""
-
-from distutils.version import LooseVersion
-
-import keras
-from keras import backend as K
+"""Tests for horovod.tensorflow.keras."""
 
 import numpy as np
 import pytest
 import tensorflow as tf
 import warnings
 
-import horovod.keras as hvd
+from distutils.version import LooseVersion
+from tensorflow import keras
+from tensorflow.python.keras import backend as K
+
+import horovod.tensorflow.keras as hvd
 
 from common import temppath
 
 
-class KerasTests(tf.test.TestCase):
+@pytest.mark.skipif(LooseVersion(tf.__version__) >= LooseVersion('2.0.0'), reason='TensorFlow v1 tests')
+class TfKerasTests(tf.test.TestCase):
     """
-    Tests for ops in horovod.keras.
+    Tests for ops in horovod.tensorflow.keras.
     """
 
     def __init__(self, *args, **kwargs):
-        super(KerasTests, self).__init__(*args, **kwargs)
+        super(TfKerasTests, self).__init__(*args, **kwargs)
         warnings.simplefilter('module')
         hvd.init()
 
         self.config = tf.ConfigProto()
         self.config.gpu_options.allow_growth = True
         self.config.gpu_options.visible_device_list = str(hvd.local_rank())
+
+    def test_train_model(self):
+        with self.test_session(config=self.config) as sess:
+            K.set_session(sess)
+
+            opt = keras.optimizers.RMSprop(lr=0.0001)
+            opt = hvd.DistributedOptimizer(opt)
+
+            model = keras.models.Sequential()
+            model.add(keras.layers.Dense(2, input_shape=(3,)))
+            model.add(keras.layers.RepeatVector(3))
+            model.add(keras.layers.ThresholdedReLU(0.5))
+            model.compile(loss=keras.losses.mean_squared_error,
+                          optimizer=opt,
+                          metrics=[keras.metrics.categorical_accuracy],
+                          sample_weight_mode='temporal')
+
+            x = np.random.random((1, 3))
+            y = np.random.random((1, 3, 3))
+
+            def generator():
+                while 1:
+                    yield (x, y)
+
+            # No assertions, we just need to verify that it doesn't hang
+            callbacks = [hvd.callbacks.BroadcastGlobalVariablesCallback(0)]
+            model.fit_generator(generator(),
+                                steps_per_epoch=10,
+                                callbacks=callbacks,
+                                epochs=0,
+                                verbose=0,
+                                workers=4,
+                                initial_epoch=1)
 
     def test_sparse_as_dense(self):
         with self.test_session(config=self.config) as sess:
@@ -53,7 +86,7 @@ class KerasTests(tf.test.TestCase):
 
             model = keras.models.Sequential()
             model.add(keras.layers.Embedding(1000, 64, input_length=10))
-            model.compile(loss=keras.losses.MSE,
+            model.compile(loss=keras.losses.mean_squared_error,
                           optimizer=opt)
 
             x = np.random.randint(1000, size=(32, 10))
@@ -183,7 +216,7 @@ class KerasTests(tf.test.TestCase):
             return model
 
         with temppath() as fname:
-            with self.test_session(config=self.config) as sess:
+            with self.session(config=self.config) as sess:
                 K.set_session(sess)
 
                 model = create_model()
@@ -196,8 +229,10 @@ class KerasTests(tf.test.TestCase):
                     model.save(fname)
 
             K.clear_session()
-            with self.test_session(config=self.config) as sess:
+            with self.session(config=self.config) as sess:
                 K.set_session(sess)
+
+                weight = np.random.random((1, 3))
 
                 if hvd.rank() == 0:
                     model = hvd.load_model(fname)
@@ -206,7 +241,7 @@ class KerasTests(tf.test.TestCase):
 
                 def generator():
                     while 1:
-                        yield (x, y)
+                        yield (x, y, weight)
 
                 if hvd.rank() == 0:
                     self.assertEqual(len(model.optimizer.weights), 5)
@@ -216,12 +251,12 @@ class KerasTests(tf.test.TestCase):
                 # No assertions, we just need to verify that it doesn't hang
                 callbacks = [hvd.callbacks.BroadcastGlobalVariablesCallback(0)]
                 model.fit_generator(generator(),
-                                    steps_per_epoch=10,
+                                    steps_per_epoch=1,
                                     callbacks=callbacks,
-                                    epochs=0,
+                                    epochs=1,
                                     verbose=0,
                                     workers=4,
-                                    initial_epoch=1)
+                                    initial_epoch=0)
 
                 self.assertEqual(len(model.optimizer.weights), 5)
 
@@ -253,23 +288,23 @@ class KerasTests(tf.test.TestCase):
             K.set_session(sess)
 
             v = 1.0 if hvd.rank() == 0 else 2.0
-            model1 = keras.models.Sequential([
-                keras.layers.Dense(2, activation='softmax')
+            model1 = tf.keras.Sequential([
+                tf.keras.layers.Dense(2, activation='softmax')
             ])
             model1.build((2, 2))
             model1.set_weights(
                 [np.array([[v,  v], [v, v]], dtype=np.float32),
                  np.array([v, v], dtype=np.float32)])
 
-            model2 = keras.models.Sequential([
-                keras.layers.Dense(2, activation='softmax')
+            model2 = tf.keras.Sequential([
+                tf.keras.layers.Dense(2, activation='softmax')
             ])
             model2.build((2, 2))
             model2.set_weights(
                 [np.array([[1.0,  2.0], [3.0, 4.0]], dtype=np.float32),
                  np.array([0.0, 0.0], dtype=np.float32)])
 
-            optimizer = keras.optimizers.Adam(0.001 * hvd.size())
+            optimizer = tf.keras.optimizers.Adam(0.001 * hvd.size())
 
             state = hvd.elastic.KerasState(model1, optimizer, batch=20 + hvd.rank(), epoch=10 + hvd.rank())
             state.sync()