+++ test/parallel/test_torch.py
@@ -20,6 +20,7 @@ from distutils.version import LooseVersion
 import inspect
 import itertools
 import os
+import platform
 import unittest
 import warnings
 import time
@@ -28,6 +29,7 @@ import json
 from collections.abc import Iterable
 
 import numpy as np
+import pytest
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
@@ -2077,6 +2079,7 @@ class TorchTests(unittest.TestCase):
             assert torch.allclose(hvd.allreduce(sync_bn.bias.grad, name='sync_bn.bias.grad'), bn.bias.grad, 1e-6)
             assert torch.allclose(hvd.allreduce(ts1.grad, name='ts1.grad'), ts2.grad, 1e-6)
 
+    @pytest.mark.skip(reason='https://github.com/horovod/horovod/issues/2330')
     def test_timeline_api(self):
         hvd.init()